BIM: move tests to own module, separate App/Gui tests, add TestArchBase
