+++ src/Mod/BIM/CMakeLists.txt
@@ -26,8 +26,6 @@ SET(Arch_SRCS
     ArchStairs.py
     ArchSpace.py
     ArchRebar.py
-    TestArch.py
-    TestArchWall.py
     ArchFrame.py
     ArchPanel.py
     ArchEquipment.py
@@ -197,6 +195,13 @@ SET(nativeifc_SRCS
     nativeifc/ifc_classification.py
 )
 
+SET(bimtests_SRCS
+    TestArch.py
+    TestArchGui.py
+    bimtests/TestArchBase.py
+    bimtests/TestArchWall.py
+)
+
 SOURCE_GROUP("" FILES ${Arch_SRCS})
 
 SET(BIMGuiIcon_SVG
@@ -215,6 +220,7 @@ ADD_CUSTOM_TARGET(BIM ALL
             ${ArchGuiIcon_SVG}
             ${importers_SRCS}
             ${bimcommands_SRCS}
+            ${bimtests_SRCS}
             ${nativeifc_SRCS}
             ${BIMGuiIcon_SVG}
 )
@@ -227,6 +233,7 @@ fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/Mod/BIM" ${Arch_SRCS})
 fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/Mod/BIM" ${Dice3DS_SRCS})
 fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/Mod/BIM" ${importers_SRCS})
 fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/Mod/BIM" ${bimcommands_SRCS})
+fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/Mod/BIM" ${bimtests_SRCS})
 fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/Mod/BIM" ${nativeifc_SRCS})
 fc_copy_sources(BIM "${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_DATADIR}/Mod/BIM" ${BIMGuiIcon_SVG})
